function for result building, functions in utils file, proper csv format in output
